Deal with multiple paths to the same peer
there's a common edge case where a device has the same fe80 address
across several mesh interfaces, this is because fe80 addresses are
derived from mac addresses and on most routers all ports are a
single physical switch.

If you have multiple connections to the same peer
via that switch failover won't work properly if you only compare ip's. So I
added interface index as a criteria as well. But this results in the creation
of one extra tunnel per peer, this tunnel is for ifidx 0 and is created when
the peer contacts us and we are unable to know what interface to contact it
back on due to the socket details provided by actix-web. Later when we hear
it's ImHere we open a proper tunnel. Previous versions simply didn't do anything
with the new ifidx and you can get away with that, but tunnels reconverge after
disruption faster if you have both sides capable of beaconing the other.

If we just stop creating tunnels for ifidx zero we break nat-punching. Cleaning ip
the ifidx zero tunnel after the creation of a more specific one is viable we should
probably do it once we have tunnel deletion code in place.